modify reconciler to use new Delta() compare
Updates the ACK runtime and type system to use the new
`pkg/compare.Delta` comparison functionality. The
`AWSResourceDescriptor.Diff()` and `AWSResourceDescriptor.Equal()`
methods were removed and replaced with a single
`AWSResourceDescriptor.Delta()` method that returns a
`*pkg/compare.Delta` struct containing the difference between two
objects under comparison.

A corresponding patch to the
github.com/aws-controllers-k8s/code-generator repository adds the
generation of the `newResourceDelta()` resource package-local function
that recurses into an object and compares the fields in the objects to
each other.

Included in this patch is an update to the `mocks` Makefile target which
now builds mocks for `k8s.io/apimachinery` interfaces that we use in a
new unit test of the update code path in the reconciler.

In order to adhere to best practices of only using exported functions in
Go module testing, I ended up exporting the
`pkg/runtime.reconciler.sync()` method as a `Sync()` method on the
`AWSResourceReconciler` interface. The unit test features a whole lot of
mock setups and definitely could use some refactoring to reduce the
number of setup mocks needed.

IMPORTANT! This is a BREAKING change to the runtime interfaces. Once
merged, we will need to cut a new Git release to signal to consumers
that this is a backwards-incompatible change and if they attempt to
regenerate their controllers with the new code and their controllers
rely on the older interfaces that uses a "diffReporter" struct, they
will generate broken code.