Add `GetHostPort` function to extract the host and port from network addresses (#29)
This patch adds a new utility function called `GetHostPort` to help
extracting the host and port from network addresses.
This function will be used in the generated `main.go` to configure
the webhook server host and port. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.