DRY up resource mock production in recon test
There was a lot of repetitive code in the pkg/runtime/reconciler_test.go
unit test file around construction of the CR mocks and the k8s
apimachinery meta object and k8s runtime object mocks.

This patch adds a utility function `resourceMocks` that returns a tuple
of the resource mock, the k8s runtime object mock, and the meta object
associated with that k8s runtime object/resource.

This allows unit tests to reduce the amount of boilerplate code
substantially and focus on the specific differences to those mocks that
impact the code behaviour under test.