Add manager filter for generation changed (#41)
Fixes: https://github.com/aws-controllers-k8s/community/issues/886

The reconciler will indefinitely requeue even when there have been no changes made to the `spec` or `status`. This is happening in both the adoption and resource reconciler types.

The root cause comes from `controller-runtime` requeue-ing the resource when we patch the `status` subresource - see [this issue](https://github.com/kubernetes-sigs/kubebuilder/issues/618) for details. This bug was introduced for the resource reconciler as part of #39 , since we now call `patchResourceStatus` on every reconcile loop.

This fix adds an event filter to each manager, with a predicate that the resource must have changed generation. The generation is not changed unless there has been a modification to the `spec`.