Use endpointResolver and add serviceEndpointsID to serviceController (#44)
Issue #, if available:
https://github.com/aws-controllers-k8s/community/issues/896

Description of changes:
If `endpointURL` is not an empty string then we will use the `endpointresolver` and we will use this `endpointURL` when the service is equal to the `serviceEndpointID` which is retrieved from the service controller. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.