Set Terminal condition on updating unmanaged (#45)
ACK uses finalizers to determine whether it is currently managing a k8s resource. We set the finalizer when creating the resource and delete it when we have deleted the underlying AWS resource. The current code path also sets the finalizer if one is not applied when attempting to update a resource. If a user were to create a resource that has the same name as an existing AWS resource, it will attempt to call `updateResource` which would set the finalizer and treat the resource as if ACK created it. This behaviour circumvents the safer adoption mechanisms put in place to prevent overriding existing resources.

This pull request ensures that any resource which does not have a finalizer, when attempting to call `updateResource`, will be marked with a terminal condition. Adopted resources are now marked with the finalizer when they are created.