Add `Status` when adopting resource (#47)
`r.kc.Create()` does not add the `Status` fields from `ReadOne` to the newly created resource. Some resources (like SageMaker's `ModelPackage`) require these resource fields as their identifier for the `ReadOne` call. It is necessary to patch the `Status` fields so the consequent update loop is able to find the resource.