add clarity to validate error message (#58)
Issue #, if available:
N/A

Description of changes:
With the removal of `--aws-account-id` the validate error message that gets propagated is less clear then it used to be. This PR aims to add more clarity to the error message that gets propagated when `ack_user_secrets` does not exist in the namespace that a given controller runs in.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.