Add `Region()` to `ResourceMetadata` struct and `AWSResourceIdentifiers` interface (#79)
Currently when a resource is created without setting a region in the
annotations, the controller uses either the namespace annotation or the
controller region flag. This causes the controller to miss manage the
resource if they are restart with a different region, or if run a second
controller using a different region flag.

This patch adds a new field to `ResourceMetadata` struct and
`AWSResourceIdentifiers` interface to store the AWS region in which
the resource is created/should be managed. This patch also updates the
`reconciler.getRegion` method to lookup for a region in the
status.ResourceMetadata before trying to get one from the
resource/namespace annotations or the controller configuration.

Addresses https://github.com/aws-controllers-k8s/community/issues/1220

This patch will also require some changes to the code-generator.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.