Patch metadata and spec when handling errors (#92)
Description of changes:
If any update hook code tried to return a requeue error, the reconciler would short circuit and never patch the metadata and spec. As a result, fields returned by the API server were never being persisted back into the K8s API - which in turn lead to the delta finding differences every loop. This is the root cause for https://github.com/aws-controllers-k8s/community/issues/1251. 

This pull request adds patching into the `HandleReconcileError` method, which sits at the end of the call stack where we handle the re-queuing errors. Therefore, even if an error is returned, as long as the error is not terminal, the spec will be patched.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.