Revert "Patch metadata and spec when handling errors (#92)" (#96)
This reverts commit de2738269396ec6e038b101745cd23c06555cc9a.

Description of changes:
Adding a patch in `HandleReconcileError` caused subsequent test failures in the EKS controller, and ultimately did not fix the issue in the RDS controller. This PR is to revert that change so that we can roll back to a known good state for any other controllers that may have been affected.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.