Update logger to ISO8601 (#98)
Description of changes:
The current logs are written in Unix nanosecond timestamp, but are awkwardly represented with scientific notation (eg. `1.6559302967712986e+09`). These timestamps are almost impossible for a human to read and because of their notation are also not suitable for log ingestion.

This pull request updates the logger to use the ISO8601 standard format (eg. `2022-06-23T23:19:35.040Z`). This should make it much clearer to developers (and end users) how much time elapsed between events. ISO8601 is a standard format for timestamps that is properly supported by log ingest tools, as well.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.