Add optional key to the FieldExport target (#100)
FieldExport generates the key name with an output structure like
`<namespace>.<FieldExport-resource-name>`. This default naming structure
creates non-conflicting names as the keys. However, many applications
expect short names as keys. This is important, especially when the
Secret/ConfigMap is mounted as files; the file names are nothing but
the key names.

This PR adds an optional attribute (.spec.to.key) to specify the key
name in the FieldExport resource. If this attribute is specified,
use this key name instead of the default.

Signed-off-by: Baiju Muthukadan <baiju.m.mail@gmail.com>

Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1410

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.