Add `inline` directive to the secret reference (#111)
The `SecretKeyReference` uses the `SecretReference` as an embedded stuct. When serializing through the apimachinery's [DefaultUnstructuredConverter](https://pkg.go.dev/k8s.io/apimachinery/pkg/runtime#pkg-variables) the json `inline` directive needs to be presented so that the fields are inline.

Issue #, if available:

Description of changes:

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.