Update reference resolution interface types (#121)
Description of changes:
Creates a new `ReferenceManager` type that re-defines `ResolveReferences` and adds a new `ClearResolvedReferences` methods to `ResourceManager`. `ResolveReferences` now returns a copy of the original `AWSResource`. `ClearResolvedReferences` also returns a copy of the original `AWSResource`.

The reconciler now resolves the references once at the beginning and passes around the resolved versions of the resource. Right before the resource is patched, in `patchMetadataAndSpec`, the resource is cleaned of references so that they are not persisted to `etcd`.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.