Set `klog` logger using the ACK runtime logger.
This patch addresses an issue where the leaderElection feature was using
`klog/v2` behind the scenes, which was set using a different logger than
the one we use in ACK. This inconsitancy caused the controller to have
two different log styles.

Fortunetly `klog/v2` allow users to set their own loggers, and more
fortunetely `go-logr` is compatible with both `klog` and `zap` the
logger we use in ACK runtime.

Setting `klog` logger results in a more consistant and smooth logging
experience.

Signed-off-by: Amine <hilalyamine@gmail.com>