Enhance namespace validation and retrieval for controller scope (#132)
This patch introduces validation for user-provided namespaces, ensuring
RFC 1123 compliance and validiy for defining the controller's scope. The
ACK controllers will now validate the comma-seperated list of namespaces
and expose a method for quering and reusing these validate namespaces.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.