Add a new flag for enabling/disabling AdoptedResource reconciler (#144)
This commit adds a new feature flag `enable-adopted-resource-reconciler` to
allow enabling or disabling the `AdoptedResource` reconciler. The reconciler
is now only started if this flag is set to true.

This lays the groundwork for a new way to adopt resources, via an adopted
resource annotation instead of using the `AdoptedResource` custom resource.

This change provides more flexibility in choosing whether users want to
use the AdoptedResource CRD or the adoption annotation to adopt
resources.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.