Update `AWSResourceManagerFactory` to include `roleARN` paramter
This patch modifies the `AWSResourceManageFactory` intgerface and
its implementation to include a `roleARN` parameter in the
`ManagerFor` function.

The addition of `roleARN` is necessary to support fine-grained
access control and multi-tenant scenarios where different IAM roles
may be used within the same AWS account and region. This change
corresponds to the updates in code-generaotr repository.