fix: Remove Finalizer from resource if create returns error (#185)
Issue [#2406#issuecomment-2860631351](https://github.com/aws-controllers-k8s/community/issues/2406#issuecomment-2860631351)

Description of changes:
The reason marking a resource as managed (putting the finalizer)
before attempting a create is a general practice in kubernetes. The main reason we 
do it is to protect against deletion protection.
If we don't put the finalizer, there is no deletion protection against the resource.

The current adoption logic expects the resource to not be managed 
(no finalizer) to trigger an adoption.
If the initial creation attempt of a resource fails due to any AWS 
error, any subsequent reconciliations attempting to adopt an existing 
resource will not succeed.

These changes set the resource as unmanaged if for any reason there is
an error during the create call, which will allow the adoption logic to 
run in subsequent reconciliations.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.