chore: Get resource from cache instead of APIServer
Currently in ACK, at the beginning of the Reconciler for each event, we
get the resource from the k8s APIServer. This is an unecessary call, as
the latest object that triggered the reconciler is already stored in the
informers cache.

With this change, we will be retrieving the resource from the cache
instead of the APIServer.