fix: report resolve references errors for adopt-or-create
Currently we have been ignoring `ResolveReferences` errors for all
adoptions and readOnly.

Ignoring the error for `adopt-or-create` is an antipattern, since we want
to ensure users are aware of these issues.

In the case of `adopt`, since the spec will fully be rewritten, we would
not need to resolve references at all.